Encrypt strings at ActiveRecord::Type::EncryptedString::Type#serialize instead of #cast
When updating a record which has an attribute typed by encrypted_string,
 `cast` is only called when the attribute actually updated.

When the attribute is not updated, just `deserialize` is called
after fetching the record from the database, and `cast` is not
called when saving that record. This causes saving the attribute
as a raw string and raise the following error:

Error:
UserTest::WhenCredentialsAreSetToEnv#test_update:
ActiveSupport::MessageEncryptor::InvalidMessage: ActiveSupport::MessageEncryptor::InvalidMessage
    test/models/user_test.rb:30:in `block in <class:WhenCredentialsAreSetToEnv>'

To avoid this error, `serialize` should be defined properly to
encrypt strings every time. `serialize` is always called before
saving those records to the database.